requirements.txt
langchain-core>=0.1.4
langchain-community>=0.0.8
langchain-openai>=0.0.2
chromadb>=0.4.18
python-dotenv>=1.0.0
openai
tiktoken
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pypdf>=3.17.1
opencc-python-reimplemented>=0.1.7
